Common build environment methods moved in build_environment.rb
